+++ Functions/randomFunctions.py
@@ -7,6 +7,9 @@ log = logging.getLogger(__name__)
 
 from random import randint
 
+import requests
+import json
+
 import Functions.message as ms										    ## Own module
 
 
@@ -29,6 +32,26 @@ def randomNumberFunction(number=11):
 def forwardedMessageFunction(message):
     forwardMessage = message['forward_from']
     return str(forwardMessage)[1:-1].replace(',', '\n')
+
+
+def shortLinkFunction(link, token):
+    params = json.dumps({'longUrl':link})
+    headers = {'content-type': 'application/json'}
+
+    url = '{0}?key={1}'.format('https://www.googleapis.com/urlshortener/v1/url', token)
+
+    response = requests.post(url, data=params, params=None, headers=headers, verify=True, timeout=5)
+    if response.ok:
+        try:
+            data = response.json()
+
+        except ValueError as e:
+            return (False, e)
+        if 'id' in data:
+            return (True, data['id'])
+
+    return (False, "Error Desconocido")
+
 
 
 log.info('RandomFunctions Module Loaded.')