+++ CHANGELOG.md
@@ -7,8 +7,14 @@ All notable changes will be documented in this file.
 - Command /newVote
 - Command /secretMessage
 - Command /anonymousMessage (Speak?¿)
-- Command /shortLink
 - Command /note
+
+## [0.0.3 - Snapshot.1] - 2018-02-14
+### Added
+- Command /shortLink
+
+### Changed
+- Tokens reading 
 
 ## [0.0.2 - Snapshot.1b] - 2018-02-13
 ### Added